+++ interfaces/kits/js/zip/napi/zlib_callback_info.cpp
@@ -24,7 +24,7 @@ namespace OHOS {
 namespace AppExecFwk {
 namespace LIBZIP {
 namespace {
-    constexpr size_t ARGS_ONE = 1;
+    constexpr uint8_t ARGS_ONE = 1;
 }
 
 void HandleEnvCleanup(void *data)
