+++ services/bundlemgr/src/installd_client.cpp
@@ -24,7 +24,7 @@
 namespace OHOS {
 namespace AppExecFwk {
 namespace {
-const int LOAD_SA_TIMEOUT_MS = 4 * 1000;
+const int16_t LOAD_SA_TIMEOUT_MS = 4 * 1000;
 } // namespace
 
 ErrCode InstalldClient::CreateBundleDir(const std::string &bundleDir)
